first cut, refactor common test suite logic
